Move rendering logic from dashboard page to CardRow component
